ai-agents/requirements.txt
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.0
web3==6.20.0
colorama==0.4.6
rich==13.7.1